{{cookiecutter.project_slug}}/application/config/urls.py
"""config URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/1.8/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  url(r'^$', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  url(r'^$', Home.as_view(), name='home')
Including another URLconf
    1. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
"""
from django.conf import settings
from django.conf.urls import include, url
from django.conf.urls.static import static
from django.contrib import admin
from wagtail.wagtailadmin import urls as wagtailadmin_urls
from wagtail.wagtailcore import urls as wagtail_urls
from wagtail.wagtaildocs import urls as wagtaildocs_urls
from wagtail.wagtailsearch import urls as wagtailsearch_urls
from wagtail.contrib.wagtailsitemaps.views import sitemap
from django.views.generic import TemplateView
from django.views.defaults import page_not_found, server_error

admin.site.site_header = 'Site Administration'

urlpatterns = static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT) + [
    url(r'^sitemap\.xml$', sitemap),
    url(r'^dadmin/', include(admin.site.urls)),
    url(r'^wadmin/', include(wagtailadmin_urls)),
    url(r'^search/', include(wagtailsearch_urls)),
    url(r'^documents/', include(wagtaildocs_urls)),
    url(r'', include(wagtail_urls)),
]

if settings.DEBUG:
    import debug_toolbar

    urlpatterns = [
                      url(r'^400/$', TemplateView.as_view(template_name='400.html')),
                      url(r'^403/$', TemplateView.as_view(template_name='403.html')),
                      url(r'^404/$', page_not_found),
                      url(r'^500/$', server_error),
                      url(r'^styleguide/$', TemplateView.as_view(template_name='core/styleguide.html')),
                      url(r'^__debug__/', include(debug_toolbar.urls)),
                  ] + urlpatterns
