{{cookiecutter.project_slug}}/application/requirements.txt
Django<1.9
django-meta
django-markup
django-storages
django-honeypot
django-analytical
django-extensions
django-svg-templatetag
django-overextends
django-debug-toolbar
django-widget-tweaks
django-flat-theme
django-flat-responsive
django-colorfield
Wand
boto
wagtail
wagtail-metadata-mixin
brotlipy
psycopg2
whitenoise